Code clean on completion of Part 5. Updated ReadMe
