+++ todo_app/templates/index.html
@@ -72,7 +72,6 @@
 
       <br><br><br>
 
-      <!-- List the todo items here -->
       </ul>
     </div>
   </div>
