+++ todo_app/templates/layout.html
@@ -17,12 +17,6 @@
       {% block content %}
       
       
-      
-      
-      
-      
-      
-      
       {% endblock %}
     </div>
