+++ todo_app/todo_item.py
@@ -1,0 +1,11 @@
+class Item:
+    def __init__(self, id, name, status='To Do'):
+        self.id = id
+        self.name = name
+        self.status = status
+
+    @classmethod
+    def from_trello_card(cls, card, list_name):
+        return cls(card['id'], card['name'], list_name)
+
+
